don't require the user to install dependencies manually
the local dependency of mocha and grunt-cli is used.

dependencies with a binary are put into the $PATH when run from npm run scripts, you can you just call them as it would have been a global. that's what we're now doing with grunt-cli.

also don't need the travis install thingy anymore.